Commit 0.0.0.10:
Renamed activity_call_destination.xml to the more appropriate
fragment_settings_call_destination.xml and changed the one usage of it in CallDestinationFragment...also noticed an unnecessary call to setContentView(R.layout.fragment_settings_call_destination), commented this out and will remove on next commit if no issues arise.

Added a cancel ImageButton to OverlayShowingService.  The cancel button flips back to search mode if the user is currently in call mode.  As of now for some reason when this button is added the service gets killed by the system as soon as the map directions open up.  It is started again easillly by leaving directions but then disappears again when directions are brought back up....possibly a resource issue.

For now I am leaving the cancelButton commented out until I can figure out a better way or a solution to the service being stopped and started when it is present.

The long term idea is that the cancel button will be invisible until the call button is pushed once at which point it will show up for 5 seconds.  A second tap on the call icon will then make the call happen.