Commit 0.0.0.11: Placed the code that creates the cancelButton into it's own method revealCancelButton() as well as the code that removes and destroys the button.
I realize it might seem more prudent to create the button once and then just show and hide it at various times. I will be trying this later but for now the creation code involves minimal processing and it seems that while created cheap phones have trouble continuing to keep the service running in memory with two buttons created on screen at the same time.