Commit 0.0.0.20: Changed public Intent getSearchOrCallIntent(boolean) back to starting the activity itself and returning void.  No need to pass these large Intent objects around if we can avoid it, just make it in place use it and destroy it.
Also Moved the code that locates the users position over to the OverLayshowingService and out of PlacePickerActivity.  Now location is pinged everytime the user clicks the first time and then if search is requested the location is ready and handed over.  So we get some computing done before the very large and sometimes lagged task of PlaceAutocomplete overlay widget coming up for the first time.