Commit 0.0.1.3: Made a bunch of Firebase additions to the app.  Yes I know they are messy as of right now and probably error prone.  This was a quick implementation to get the ball rolling for gathering and storing data on each individual user in the Firebase realtime database.
OverlayShowingService.java creates a new UUID for the user if one has not been made yet.  It is stored in QuieryPreferences PreferenceManager as two long primitives representing the Most and Least signigicant bits of the 128 bit UUID.

QueryPreferences.java stores and returns the UUID when setUUID(Context, UUID) or getUUID(Context) are called respectively.  No work is needed to be done to take apart or reassemble the UUIDs Most and Least Significant bits, that is all done internally here.

MyPlace is a highly stripped fown version of the Google Place object that holds just the data we are interested in with regards to the users place lookup.  To the level that it doesn't even extend Place, it is just a datastore object.

PlacePickerActivity now also gets the users UUID so that it can push the data of the Place being returned from the AutocompletePlacePicker dialog to the database filed under the specific users UUID.

The database is currently structured so that each unique user has a list of times at which they were located at a specific latitude and longitude (current_location), times when they searched for a Place, and times when they at least attempted to make a phone call to the place (phone_call_attempt).