Commit 0.0.1.4:
MyPlace.java: Removed the Date timeOfCreation member variable.  All timestamps are now completely created, stored, and retrieved as a single long representing the epoch time, millis since 01-01-1970.  and MyPlace doesn't really need a time-stamp member anyways.

CallDestinationFragment.java, mostly cleanup, removed some uneeded commented out lines.  Switched it up so that everytime the settings are brought up the app starts the service if it is supposed to be started already and isn't for any reason at all.  changed setIsOn(boolean isOn) to use isOn everywhere instead of mIsServiceOn, which is set = to isOn at the top but it is more appropiate to stay with isOn in case in the future mIsServiceOn is not set = at the top or any other things.

OverlayShowingService.java: now when the service is started it checks to see if a UUID has been assigned to this particular installation of the app and if it hasn't it generates a random one and stores it as this installations UUID.  the UUID as well as the epoch time stored as a long are used to write and reference all data within the database that gets stored.

PlacePickerActivity.java: basically the same as OverlayShowingService but just doesn't make a new UUID if one is not made, instead it stores everything under the guest UUID of (0,0).  Probably should fix this...

QueryPreferences.java: capitalized UUID