0.1.0.0: The same as 0.0.1.6. but with new and shiny Fabric Crashlytics kit installed.
First commit after the mystery transition.  Project had an unsolvable error that caused a crash immediately on launch when integrating crashlytics.  The error occured at Fabric.with(this, new crashlytics()); and was primarily shown as "io.fabric.sdk.android.services.concurrency.UnmetDependencyException"  I never figured out the cause.  I started a new project and simply moved all of the source code and configuration over and everything magically worked.  So here we are at 0.1.0.0